+++ graph/shared/src/test/ch/scorpion/jabbah/graph/CanvasMockBuilder.kt
@@ -28,11 +28,11 @@ class CanvasMockBuilder {
         every { canvas.addPropertyChangeListener(capture(propertyChangeSlot)) } calls  {
             propertyChangeListeners.add(propertyChangeSlot.get())
         }
-        withDevicePixelRatio(1)
+        withDevicePixelRatio(1.0)
         withDimension(Dimension2D(0, 0))
     }
 
-    fun withDevicePixelRatio(devicePixelRatio: Int): CanvasMockBuilder {
+    fun withDevicePixelRatio(devicePixelRatio: Double): CanvasMockBuilder {
         every { canvas.devicePixelRatio } returns devicePixelRatio
         return this
     }
